puteuuuuuuuh on est sur la fin puuuuteuh

+++ server/sources/init/init_globals.c
@@ -6,7 +6,7 @@
 /*   By: nsampre <nsampre@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/01 02:57:27 by nsampre           #+#    #+#             */
-/*   Updated: 2017/12/01 02:57:27 by nsampre          ###   ########.fr       */
+/*   Updated: 2017/12/01 03:24:21 by nsampre          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ void	init_globals(t_env *e)
 	struct timeval t;
 
 	g_e = e;
-	g_buffer = (t_vector *)ft_memalloc(sizeof(t_vector) * F_WIDTH * F_HEIGHT);
+	g_buffer = (t_vector *)ft_memalloc(
+							sizeof(t_vector) * (int)F_WIDTH * (int)F_HEIGHT);
 	gettimeofday(&t, NULL);
 	srand(t.tv_usec);
 }
-
